shared services

+++ src/Guide.Application/Features/AttractionImages/Commands/DeleteAttractionImage/DeleteAttractionImageCommandHandler.cs
@@ -9,24 +9,30 @@ public class DeleteAttractionImageCommandHandler(GuideDbContext dbContext)
 {
     public async Task Handle(DeleteAttractionImageCommand request, CancellationToken cancellationToken)
     {
-        var path = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "wwwroot", "uploads", request.Image.Path);
+        var image = await dbContext.AttractionImages
+            .AsNoTracking()
+            .SingleOrDefaultAsync(x => x.Id == request.Id, cancellationToken);
+
+        if (image == null) return;
+
+        var path = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "wwwroot", "uploads", image.Path);
         if (File.Exists(path))
         {
             File.Delete(path);
         }
 
         await dbContext.AttractionImages
-            .Where(x => x.Id == request.Image.Id)
+            .Where(x => x.Id == image.Id)
             .ExecuteDeleteAsync(cancellationToken);
 
-        if (request.Image.IsMain)
+        if (image.IsMain)
         {
             await dbContext.AttractionImages
-                .Where(x => x.AttractionId == request.Image.AttractionId)
+                .Where(x => x.AttractionId == image.AttractionId)
                 .ExecuteUpdateAsync(x => x.SetProperty(x => x.IsMain, false), cancellationToken);
 
             var mainImage = await dbContext.AttractionImages
-                .Where(x => x.AttractionId == request.Image.AttractionId)
+                .Where(x => x.AttractionId == image.AttractionId)
                 .FirstOrDefaultAsync(cancellationToken);
 
             if (mainImage != null)
